requirements.txt
nuplan-devkit @ git+https://github.com/motional/nuplan-devkit/@nuplan-devkit-v1.2

# nuplan requirements
aioboto3
aiofiles
bokeh
casadi
control
Fiona
geopandas
guppy3
joblib
matplotlib
nest_asyncio
numpy
opencv-python
pandas
Pillow
psutil
pyarrow
pyinstrument
pyogrio
pyquaternion
pytest
rasterio
ray
retry
rtree
scipy
selenium
setuptools
shapely==2.1.1
SQLAlchemy==1.4.27
sympy
tornado
tqdm
ujson
notebook
pre-commit
cachetools

# hydra
hydra_colorlog
hydra-core

# d123
lxml
trimesh
viser
gym==0.17.2
gymnasium==0.26.3

# torch & lighting
torch==2.6.0
torchvision==0.21.0
lightning
tensorboard
protobuf==4.25.3
pre-commit
